requirements.txt
Flask==2.2
Flask-SQLAlchemy==3.0.0
SQLAlchemy==1.4.30
pymysql==1.0.2
Werkzeug==2.2.2
python-dotenv==0.19.2